// src/apis/generated/model/procedureDocumentDetailRes.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * 동행 API
 * 동행 서비스 API 입니다.
 * OpenAPI spec version: v1
 */
import type { ProcedureDocumentDetailResDocumentChannelType } from "./procedureDocumentDetailResDocumentChannelType";
import type { ProcedureDocumentDetailResDocumentType } from "./procedureDocumentDetailResDocumentType";

export interface ProcedureDocumentDetailRes {
  procedureDocumentId?: number;
  userDocumentChecklistId?: number;
  documentType?: ProcedureDocumentDetailResDocumentType;
  documentChannelType?: ProcedureDocumentDetailResDocumentChannelType;
  documentName?: string;
  documentLocation?: string;
  documentLink?: string;
  description?: string;
  checked?: boolean;
}